Not an ideal solution.
But what are you going to do, it's Bash? Associating one text-based thing
with another text-based thing is really hard to do without associative
arrays, and associative arrays appear to not be universally supported.

While this isn't the most elegant solution, I think that you just have
to deal with it for this specific application.